test(API): add tests covering ReferenceDataController constructor
Add test methods covering the constructor for ReferenceDataController. May add
tests to cover endpoints later, but methods are currently very simple and will not
benefit much from unit tests. Will use Postman covering integration tests to address
testing endpoints.
